fix:/UpdateDeviceGroupSenderEvent接口健壮处理,屏蔽空数据列表情况

+++ src/main/java/cn/shmedo/monitor/monibotbaseapi/listener/UpdateEventListener.java
@@ -40,12 +40,14 @@ public class UpdateEventListener {
      * @see UpdateDeviceGroupSenderDto
      */
     private void clearDeviceWarnLog(final List<UpdateDeviceGroupSenderEventParam> dataList) {
-        LambdaQueryWrapper<TbDeviceWarnLog> wrapper = new LambdaQueryWrapper<TbDeviceWarnLog>().nested(wr ->
-                dataList.forEach(data -> wr.or(u -> {
-                    u.eq(TbDeviceWarnLog::getDeviceToken, data.getDeviceToken());
-                    Optional.ofNullable(data.getProjectIDList()).filter(CollUtil::isNotEmpty)
-                            .ifPresent(w -> u.notIn(TbDeviceWarnLog::getProjectID, w));
-                })));
-        tbDeviceWarnLogMapper.delete(wrapper);
+        if (CollUtil.isNotEmpty(dataList)) {
+            LambdaQueryWrapper<TbDeviceWarnLog> wrapper = new LambdaQueryWrapper<TbDeviceWarnLog>().nested(wr ->
+                    dataList.forEach(data -> wr.or(u -> {
+                        u.eq(TbDeviceWarnLog::getDeviceToken, data.getDeviceToken());
+                        Optional.ofNullable(data.getProjectIDList()).filter(CollUtil::isNotEmpty)
+                                .ifPresent(w -> u.notIn(TbDeviceWarnLog::getProjectID, w));
+                    })));
+            tbDeviceWarnLogMapper.delete(wrapper);
+        }
     }
 }
